refactor(client): shifts and new subset
- add page folder into components
- move client index into app
- modify webpack config src directory to new app path
- rename godirectoryinput as directoryinput
- move gosearchinput into components as it is shared
- add go logo graphic folder into assets
