+++ src/client/app/assets/go-logo-graphics/go-qr-logo.png
