+++ src/client/app/components/BaseLayout/BaseLayoutHeader.jsx
@@ -12,10 +12,10 @@ import {
   useTheme,
 } from '@material-ui/core'
 import i18next from 'i18next'
-import GoLogo from '../../assets/go-main-logo.svg'
-import GoLogoLight from '../../assets/go-main-logo-light.svg'
-import GoLogoMini from '../../assets/go-main-logo-mini.svg'
-import GoLogoMiniLight from '../../assets/go-main-logo-mini-light.svg'
+import GoLogo from '../../assets/go-logo-graphics/go-main-logo.svg'
+import GoLogoLight from '../../assets/go-logo-graphics/go-main-logo-light.svg'
+import GoLogoMini from '../../assets/go-logo-graphics/go-main-logo-mini.svg'
+import GoLogoMiniLight from '../../assets/go-logo-graphics/go-main-logo-mini-light.svg'
 import helpIcon from '../../assets/help-icon.svg'
 import loginActions from '../../../login/actions'
 import Section from '../Section'
