+++ src/client/app/components/pages/RootPage/index.jsx
@@ -4,22 +4,22 @@ import { Provider } from 'react-redux'
 import { HashRouter, Route, Switch } from 'react-router-dom'
 import { MuiThemeProvider } from '@material-ui/core/styles'
 
-import PrivateRoute from '../PrivateRoute'
-import HomePage from '../../../home'
-import LoginPage from '../../../login'
-import UserPage from '../../../user'
+import PrivateRoute from '../../PrivateRoute'
+import HomePage from '../../../../home'
+import LoginPage from '../../../../login'
+import UserPage from '../../../../user'
 import NotFoundPage from '../NotFoundPage'
-import SearchPage from '../../../search'
-import DirectoryPage from '../../../directory'
-import MessageSnackbar from '../MessageSnackbar'
+import SearchPage from '../../../../search'
+import DirectoryPage from '../../../../directory'
+import MessageSnackbar from '../../MessageSnackbar'
 import ScrollToTop from './ScrollToTop'
 
 // Add extra favicons to webpack bundle
-import '../../assets/favicon/android-chrome-192x192.png'
-import '../../assets/favicon/android-chrome-512x512.png'
-import '../../assets/favicon/apple-touch-icon.png'
-import '../../assets/favicon/favicon-16x16.png'
-import '../../assets/favicon/favicon-32x32.png'
+import '../../../assets/favicon/android-chrome-192x192.png'
+import '../../../assets/favicon/android-chrome-512x512.png'
+import '../../../assets/favicon/apple-touch-icon.png'
+import '../../../assets/favicon/favicon-16x16.png'
+import '../../../assets/favicon/favicon-32x32.png'
 
 import {
   DIRECTORY_PAGE,
@@ -28,8 +28,8 @@ import {
   NOT_FOUND_PAGE,
   SEARCH_PAGE,
   USER_PAGE,
-} from '../../util/types'
-import theme from '../../theme'
+} from '../../../util/types'
+import theme from '../../../theme'
 
 const Root = ({ store }) => (
   <Provider store={store}>
