+++ src/client/app/reducers/index.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux'
 import { login } from '../../login/reducers'
 import user from '../../user/reducers'
-import root from '../components/RootPage/reducers'
+import root from '../components/pages/RootPage/reducers'
 import home from '../../home/reducers'
 import search from '../../search/reducers'
 import directory from '../../directory/reducers'
