+++ src/client/app/util/download.js
@@ -1,6 +1,6 @@
 import { saveAs } from 'file-saver'
 import * as Sentry from '@sentry/browser'
-import rootActions from '../components/RootPage/actions'
+import rootActions from '../components/pages/RootPage/actions'
 import userActions from '../../user/actions'
 import { useIsIE } from '../components/BaseLayout/util/ie'
 import { GAEvent } from './ga'
