+++ src/client/directory/DirectoryResults/DirectoryTable/DirectoryTableRow/index.tsx
@@ -14,8 +14,8 @@ import {
 import { UrlTypePublic } from '../../../reducers/types'
 import useAppMargins from '../../../../app/components/AppMargins/appMargins'
 import personIcon from '../../../assets/person-icon.svg'
-import { SetSuccessMessageAction } from '../../../../app/components/RootPage/actions/types'
-import rootActions from '../../../../app/components/RootPage/actions'
+import { SetSuccessMessageAction } from '../../../../app/components/pages/RootPage/actions/types'
+import rootActions from '../../../../app/components/pages/RootPage/actions'
 import DirectoryFileIcon from '../../../widgets/DirectoryFileIcon'
 import DirectoryUrlIcon from '../../../widgets/DirectoryUrlIcon'
